refactored places support.
Replaced `-i` option with `-c` option.  This allows you to specify which
columns get imported and optionally map those columns to property names
using a colon character.

Added `places-update` type, this allows you to do updates on places.
Mainly this is useful when setting the contact/person on a place, an
extra request is needed to define that relation because it's circular.

Removed the `-t` flag because type is required, made it a command line
parameter.

Moved validation call into row handler.

Issue: https://github.com/medic/medic-webapp/issues/2568